feat: separa popup de criaçao de novos projetos em um componente
