+++ src/frontend/src/components/Sidebar.jsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from 'react'
+import React, { useState } from 'react'
 import styled from 'styled-components'
 import { IoIosAdd } from "react-icons/io"
 import { FaRegUserCircle, FaFolder } from "react-icons/fa"
@@ -6,6 +6,7 @@ import { IoExitOutline } from "react-icons/io5"
 import logo from '../logo.svg'
 import { useProject } from '../contexts/ProjectContext'
 import { useNavigate } from 'react-router-dom'
+import NovoProjetoPopup from './NovoProjetoPopup'
 
 const Container = styled.div`
   width: 18vw;
@@ -132,59 +133,6 @@ const Perfil = styled.div`
   }
 `
 
-const Popup = styled.div`
-  background-color: #3D80A3;
-  width: 15vw;
-  padding: 2rem;
-  border: 5px solid #2E2E2E;
-  border-radius: 20px;
-  color: #fff;
-
-  display: flex;
-  flex-direction: column;
-
-  position: fixed;
-  top: 21%;
-  left: 1%;
-
-  h3 {
-    margin: .5rem 0;
-  }
-
-  input {
-    background-color: #D7D7D7;
-    border: none;
-    width: 90%;
-  }
-
-  #nome-projeto {
-    background-color: #BDE0EE;
-    height: 2rem;
-    border-radius: 15px;
-  }
-
-  #nome-contratante, #data {
-    background-color: #97C6D9;
-    height: 1.5rem;
-    border-radius: 15px;
-  }
-
-  button {
-    border: 5px solid #629EBC;
-    border-radius: 10px;
-    padding: .3rem;
-    background-color: #BDE0EE;
-    width: 70%;
-
-    font-size: 28px;
-    color: 0A3B4E;
-  }
-
-  button:hover {
-    background-color: #97C6D9; 
-    cursor: pointer; 
-  }
-`
 
 export default function Sidebar(props) {
   const [showPopup, setShowPopup] = useState(false);
@@ -195,20 +143,6 @@ export default function Sidebar(props) {
     setShowPopup(!showPopup);
   };
 
-  const nameRef = useRef();
-  const contractorRef = useRef();
-  const dateRef = useRef();
-
-
-  const handleConclude = () => {
-    setProject({
-      name: nameRef.current.value,
-      contractor: contractorRef.current.value,
-      date: dateRef.current.value
-    });
-
-    togglePopup();
-  };
 
   const handleClick = () => {
     navigate("/projetos");
@@ -224,22 +158,7 @@ export default function Sidebar(props) {
       </BntMaior>
 
       {showPopup && (
-        <Popup>
-          <div>
-            <h3>Nome do projeto</h3>
-            <input type='text' ref={nameRef} id='nome-projeto' />
-
-            <h3>Nome do contratante</h3>
-            <input type='text' ref={contractorRef} id='nome-contratante' />
-
-            <h3> Data</h3>
-            <input type='date' ref={dateRef} id='data' />
-
-            <div style={{ width: '100%', display: 'flex', flexDirection: 'column', alignItems: 'center', marginTop: '2rem' }}>
-              <button onClick={handleConclude}>Enviar</button>
-            </div>
-          </div>
-        </Popup>
+        <NovoProjetoPopup onClose={togglePopup} onSubmit={(data) => setProject(data)} />
       )}
 
       <section>
